+++ laktory/models/datasources/eventdatasource.py
@@ -1,5 +1,7 @@
+from pydantic import model_validator
 from laktory.spark import DataFrame
 from typing import Literal
+from typing import Any
 
 from laktory.models.dataeventheader import DataEventHeader
 from laktory.models.datasources.basedatasource import BaseDataSource
@@ -25,6 +27,8 @@ class EventDataSource(BaseDataSource, DataEventHeader):
     type: Literal[TYPES] = "STORAGE_EVENTS"
     fmt: Literal[FORMATS] = "JSON"
     multiline: bool = False
+    header: bool = True
+    read_options: dict[str, str] = {}
 
     # ----------------------------------------------------------------------- #
     # Readers                                                                 #
@@ -33,29 +37,42 @@ class EventDataSource(BaseDataSource, DataEventHeader):
     def _read_storage(self, spark) -> DataFrame:
         if self.read_as_stream:
             logger.info(f"Reading {self.event_root} as stream")
-            df = (
-                spark.readStream.format("cloudFiles")
-                .option("multiLine", self.multiline)
-                .option("mergeSchema", True)
-                .option("recursiveFileLookup", True)
+
+            # Set reader
+            reader = (
+                spark.readStream
+                .format("cloudFiles")
                 .option("cloudFiles.format", self.fmt)
                 .option("cloudFiles.schemaLocation", self.event_root)
                 .option("cloudFiles.inferColumnTypes", True)
                 .option("cloudFiles.schemaEvolutionMode", "addNewColumns")
                 .option("cloudFiles.allowOverwrites", True)
-                .load(self.event_root)
             )
+
         else:
             logger.info(f"Reading {self.event_root} as static")
-            df = (
-                spark.read.option("multiLine", self.multiline)
-                .option("mergeSchema", True)
-                .option("recursiveFileLookup", True)
+
+            # Set reader
+            reader = (
+                spark.read
                 .format(self.fmt)
-                .load(self.event_root)
             )
-            # Not supported by UC
-            # .withColumn("file", F.input_file_name())
+
+        reader = (
+            reader
+            .option("multiLine", self.multiline)  # only apply to JSON format
+            .option("mergeSchema", True)
+            .option("recursiveFileLookup", True)
+            .option("header", self.header)  # only apply to CSV format
+        )
+        if self.read_options:
+            reader = reader.options(**self.read_options)
+
+        # Load
+        df = reader.load(self.event_root)
+
+        # Not supported by UC
+        # .withColumn("file", F.input_file_name())
 
         return df